+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add vertical pod autoscaler support.
+- Add `appversionlabel` resource to update version labels for optional app CRs.
 
 ## [3.4.1] - 2020-12-03